Complit UserStatistic DataBase

+++ domain/src/main/java/com/BySandS/testsandquizes/domain/allData/models/mainMenu/GameCompletionPercentageModel.kt
@@ -1,5 +1,5 @@
 package com.BySandS.testsandquizes.domain.allData.models.mainMenu
 
 class GameCompletionPercentageModel (
-    var percentGameCompletion: Int
+    var percentGameCompletion: Double,
 )
